requirements.txt
boto3
flask
flask_jwt_extended
flask-restful
python-dotenv
requests
selenium